requirements.txt
# fastapi==0.95.2
# uvicorn==0.22.0
# sqlalchemy==2.0.20
# psycopg2-binary==2.9.7
# python-dotenv==1.0.0
# alembic==1.11.1
# pydantic==2.0.0


fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
alembic==1.11.1
pydantic

